src/test/java/sonia/scm/tagprotection/TagProtectionIndexLinkEnricherTest.java
/*
 * Copyright (c) 2020 - present Cloudogu GmbH
 *
 * This program is free software: you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License as published by the Free
 * Software Foundation, version 3.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 */

package sonia.scm.tagprotection;

import com.github.sdorra.shiro.ShiroRule;
import com.github.sdorra.shiro.SubjectAware;
import com.google.inject.util.Providers;
import jakarta.inject.Provider;
import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;
import sonia.scm.api.v2.resources.HalAppender;
import sonia.scm.api.v2.resources.HalEnricherContext;
import sonia.scm.api.v2.resources.ScmPathInfoStore;

import java.net.URI;

import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.verifyNoInteractions;

@RunWith(MockitoJUnitRunner.class)
public class TagProtectionIndexLinkEnricherTest {

    @Rule
    public ShiroRule shiro = new ShiroRule();

    @Mock
    private HalAppender appender;

    private TagProtectionIndexLinkEnricher enricher;

    @Before
    public void setUp() {
        ScmPathInfoStore scmPathInfoStore = new ScmPathInfoStore();
        scmPathInfoStore.set(() -> URI.create("https://scm-manager.org/scm/api/"));
        Provider<ScmPathInfoStore> scmPathInfoStoreProvider = Providers.of(scmPathInfoStore);
        enricher = new TagProtectionIndexLinkEnricher(scmPathInfoStoreProvider);
    }

    @SubjectAware(
            username = "trillian",
            password = "secret",
            configuration = "classpath:sonia/scm/tagprotection/shiro.ini"
    )
    @Test
    public void shouldEnrich() {
        enricher.enrich(HalEnricherContext.of(), appender);
        verify(appender).appendLink("tagProtection", "https://scm-manager.org/scm/api/v2/config/tagprotection/");
    }

    @SubjectAware(
            username = "unpriv",
            password = "secret",
            configuration = "classpath:sonia/scm/tagprotection/shiro.ini"
    )
    @Test
    public void shouldNotEnrichIfNotPermitted() {
        enricher.enrich(HalEnricherContext.of(), appender);
        verifyNoInteractions(appender);
    }
}
